stert of add groups,bag fixing

+++ src/client/client tk.py
@@ -53,14 +53,71 @@ listbox.bind('<<ListboxSelect>>', on_select)
 
 
 #=============================================================================
+'''
+class move_dialog:
+    def __init__(self, root,filebox):
+
+        self.root = root
+        top = self.top = Toplevel(root)
+
+        eror_string = StringVar(value="")
+        self.listbox=Listbox(top,  width = 30,height=10, font='Arial 12')
+        def on_select(event):
+            global  file_selected
+            try:
+                index= listbox.curselection()[0]
+                file_selected=listbox.get(index).split(" ")
+            except:pass
+
+
+        new_f=partial(self.ok,eror_string,filebox)
+        b = Button(top, text="OK", command=new_f)
+        b.pack(pady=5)
+        self.path=["","",""]
+        self.file_selected=""
+
+
+
+    def ok(self,eror_string,filebox):
+        global username
+        global pasward
+        global conn
+        global path
+
+        username2 = self.e1.get()
+        pasward2 = self.e2.get()
+        s=check_pasward(username2,pasward2,conn)
+
+        if s=="there was a conection eror":
+            eror_string.set("there was a conection eror")
+        elif s==False:
+            eror_string.set("the username or the pasward is incorect")
+        else:
+            username=username2+''
+            pasward=pasward2+''
+            path='/'+username
+            write_file_list(get_file_list(username,path,conn), filebox)
+            self.top.destroy()
+
+class change_permition:
+    def __init__(self, root,filebox):
+
+        self.root = root
+        top = self.top = Toplevel(root)
+
+        eror_string = StringVar(value="")
+        self.listbox=Listbox(top,  width = 30,height=10, font='Arial 12')
+
+'''
 path_string = StringVar(value="")
 path_label=Label(root ,textvariable=path_string, font='Arial 13')
 
 buttons_frame=Frame(root)
 b_font='Arial 12'
 def back2(filebox,path_string):
+    global username
     global path
-    if len(path.split("/"))>1:
+    if path.split("/")[-1]!=username:
         path="/".join(path.split("/")[:-1])
         write_file_list(get_file_list(username,path,conn), filebox)
         path_string.set(path)
@@ -70,12 +127,13 @@ def delete2(filebox):
     global path
     global file_selected
     global conn
-    if mb.askyesno('Verify', 'you sure that you want to delete '+file_selected[0]+"?"):
-        q=delete_file(username,path+'/'+file_selected[0],conn)
-        if q=="the file doesnt exists":mb.showerror("Error", "Error message")
-        elif q=="the user isn't have permishion to do it":mb.showerror("Error", "you dont have pemittion to do it")
-        elif q=="ok":
-            write_file_list(get_file_list(username,path,conn), filebox)
+    if file_selected!=["","",""]:
+        if mb.askyesno('Verify', 'you sure that you want to delete '+file_selected[0]+"?"):
+            q=delete_file(username,path+'/'+file_selected[0],conn)
+            if q=="the file doesnt exists":mb.showerror("Error", "Error message")
+            elif q=="the user isn't have permishion to do it":mb.showerror("Error", "you dont have pemittion to do it")
+            elif q=="ok":
+                write_file_list(get_file_list(username,path,conn), filebox)
 delete=partial(delete2,listbox)
 def uploud():pass
 
@@ -93,22 +151,23 @@ def rename2(root,filebox):
     global path
     global file_selected
     global conn
-    name= sd.askstring("Input", "new name",parent=root)
-    q=rename_file(username,str(path+'/'+file_selected[0]),name,conn)
-    file_selected[0]=name
-    if q=="the file doesnt exists":mb.showerror("Error", "Error message")
-    elif q=="the user isn't have permishion to do it":mb.showerror("Error", "you dont have pemittion to do it")
-    elif q=="ok":
-        write_file_list(get_file_list(username,path,conn), filebox)
+    if file_selected!=["","",""]:
+        name= sd.askstring("Input", "new name",parent=root)
+        q=rename_file(username,str(path+'/'+file_selected[0]),name,conn)
+        file_selected[0]=name
+        if q=="the file doesnt exists":mb.showerror("Error", "Error message")
+        elif q=="the user isn't have permishion to do it":mb.showerror("Error", "you dont have pemittion to do it")
+        elif q=="ok":
+            write_file_list(get_file_list(username,path,conn), filebox)
 rename=partial(rename2,root,listbox)
 
-def replace():pass
+def move():pass
 
 open_b=Button(buttons_frame, text = "open",command=open_dir, font=b_font)
 open_b.pack(side = LEFT, expand = 1)
 
 
-for f_name in ['back','delete','uploud','rename','replace']:
+for f_name in ['back','delete','uploud','rename','move']:
     Button(buttons_frame, text = f_name,command=eval(f_name), font=b_font).pack(side = LEFT, expand = 1)
 
 buttons_frame.pack(side=BOTTOM)
@@ -206,15 +265,20 @@ class sign_up_dialog:
         else:
             self.top.destroy()
 
+
+
 frame=Frame(root)
 hello_text=StringVar(value="   hello   \n")
 lb2 = Label(frame ,textvariable=hello_text,bg="#b0b0e0", font='Arial 15').grid(row=0,column=0)
 
 def log_in(hello_text,lisbox,path_string):
+    global path
     d = log_in_dialog(root,listbox)
     d.root.wait_window(d.top)
     hello_text.set("   hello   \n"+username)
-    path_string.set(path)
+    print(file_selected)
+    if path!="":
+        path_string.set(path)
 log_in2=partial(log_in,hello_text,listbox,path_string)
 log_in_b=Button(frame, text = "log in",command=log_in2 ,  width = 10, font='Arial 12') .grid(row=1,column=0)#.pack(side = TOP, fill = X,anchor='ne')
 
@@ -222,7 +286,7 @@ def log_out(hello_text,path_string):
     global username
     global pasward
     global path
-    if not username=='':
+    if username!='':
         if mb.askyesno('Verify', 'you sure that you want to log out?'):
             username=""
             pasward=""
